+++ src/test/java/com/scriptbasic/utility/TestCastUtility.java
@@ -1,6 +1,3 @@
-/**
- *
- */
 package com.scriptbasic.utility;
 
 import com.scriptbasic.errors.BasicInterpreterInternalError;
@@ -20,16 +17,16 @@ public class TestCastUtility {
     public void test() throws Exception {
         assertEquals(CastUtility.cast(13.5, Byte.class), (byte) 13);
         assertEquals(CastUtility.cast(13.5, Short.class), (short) 13);
-        assertEquals(CastUtility.cast(13.5, Integer.class), (int) 13);
+        assertEquals(CastUtility.cast(13.5, Integer.class), 13);
         assertEquals(CastUtility.cast(13.5, Long.class), (long) 13);
         assertEquals(CastUtility.cast(13.5, Float.class), (float) 13.5);
-        assertEquals(CastUtility.cast(13.5, Double.class), (double) 13.5);
+        assertEquals(CastUtility.cast(13.5, Double.class), 13.5);
         assertEquals(CastUtility.cast(13.5, byte.class), (byte) 13);
         assertEquals(CastUtility.cast(13.5, short.class), (short) 13);
-        assertEquals(CastUtility.cast(13.5, int.class), (int) 13);
+        assertEquals(CastUtility.cast(13.5, int.class), 13);
         assertEquals(CastUtility.cast(13.5, long.class), (long) 13);
         assertEquals(CastUtility.cast(13.5, float.class), (float) 13.5);
-        assertEquals(CastUtility.cast(13.5, double.class), (double) 13.5);
+        assertEquals(CastUtility.cast(13.5, double.class), 13.5);
 
         assertEquals(CastUtility.cast("DEADBEEF", Byte.class), "DEADBEEF");
         assertEquals(CastUtility.cast("DEADBEEF", Short.class), "DEADBEEF");
